@@ -1,5 +1,6 @@
-import VueDraggableResizable from '@/components/vue-draggable-resizable'
+import VueDraggableResizable from '@/components/vue-draggable-resizable.vue'
 import { mount } from '@vue/test-utils'
+import { describe, it, expect, afterEach } from 'vitest'
 import syn from 'syn'
 import div from '../div'
 
@@ -157,5 +158,5 @@ describe('events', function () {
     })
   })
 
-  afterEach(() => wrapper.destroy())
+  afterEach(() => wrapper.unmount())
 })
